Move formatAmountHuman to lib and merge it with identical implementations
